# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于管理车位、用户、租赁的接口服务。",
        "title": "停车系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "技术支持",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/parking/entry": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "车辆入场时登记车牌号，开始计费",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking"
                ],
                "summary": "车辆入场登记",
                "parameters": [
                    {
                        "description": "入场信息",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.EntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "入场记录",
                        "schema": {
                            "$ref": "#/definitions/controllers.RecordResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parking/exit/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "车辆出场时结算停车费用，返回停车记录和费用信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking"
                ],
                "summary": "车辆出场结算",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "停车记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "出场结算记录",
                        "schema": {
                            "$ref": "#/definitions/controllers.RecordResponse"
                        }
                    },
                    "400": {
                        "description": "无效的ID参数",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parking/rent": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户将自己的车位发布出租，需指定车位ID、出租价格及出租天数。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking"
                ],
                "summary": "出租车位",
                "parameters": [
                    {
                        "description": "出租车位请求体",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功返回租赁信息",
                        "schema": {
                            "$ref": "#/definitions/controllers.LeaseResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/parking/spots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有车位的详细列表，包括类型、状态和收费标准",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking"
                ],
                "summary": "获取车位列表",
                "responses": {
                    "200": {
                        "description": "返回车位列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ParkingSpot"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新增一个车位，指定车位类型和收费标准",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking"
                ],
                "summary": "创建车位",
                "parameters": [
                    {
                        "description": "车位信息",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateSpotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功的车位信息",
                        "schema": {
                            "$ref": "#/definitions/controllers.ParkingSpotResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reports/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据指定天数获取每日停车收入和车位使用统计，默认返回最近7天的数据。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "获取日报表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 7,
                        "description": "查询天数，默认为7天",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "日报表数据",
                        "schema": {
                            "$ref": "#/definitions/controllers.DailyReportResponse"
                        }
                    },
                    "400": {
                        "description": "无效的查询参数",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vehicles": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将车辆绑定到当前用户账户，支持填写车牌号、品牌及车型信息。业主绑定车辆（可绑定多个）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle"
                ],
                "summary": "绑定车辆",
                "parameters": [
                    {
                        "description": "绑定车辆请求体",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.BindVehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定成功返回车辆信息",
                        "schema": {
                            "$ref": "#/definitions/controllers.VehicleResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.BindVehicleRequest": {
            "type": "object",
            "required": [
                "license"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateSpotRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "hourly_rate": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/models.ParkingType"
                }
            }
        },
        "controllers.DailyReportResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "permanent_count": {
                    "type": "integer"
                },
                "short_term_count": {
                    "type": "integer"
                },
                "temporary_count": {
                    "type": "integer"
                },
                "total_income": {
                    "type": "number"
                }
            }
        },
        "controllers.EntryRequest": {
            "type": "object",
            "required": [
                "license"
            ],
            "properties": {
                "license": {
                    "type": "string"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "controllers.LeaseResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "结束日期",
                    "type": "string"
                },
                "id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "spot_id": {
                    "description": "车位ID",
                    "type": "integer"
                },
                "start_date": {
                    "description": "起始日期（格式：YYYY-MM-DD）",
                    "type": "string"
                },
                "status": {
                    "description": "当前状态",
                    "type": "string"
                },
                "total": {
                    "description": "总金额",
                    "type": "number"
                }
            }
        },
...
        "controllers.ParkingSpotResponse": {
            "type": "object",
            "properties": {
                "hourly_rate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "controllers.RecordResponse": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "entry_time": {
                    "type": "string"
                },
                "exit_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license": {
                    "type": "string"
                },
                "spot_id": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.RentRequest": {
            "type": "object",
            "required": [
                "days",
                "rate",
                "spot_id"
            ],
            "properties": {
                "days": {
                    "type": "integer",
                    "minimum": 1
                },
                "rate": {
                    "type": "number"
                },
                "spot_id": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.VehicleResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "license": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                }
            }
        },
        "models.ParkingSpot": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "description": "添加过期时间字段",
                    "type": "string"
                },
                "hourlyRate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "license": {
                    "type": "string"
                },
                "monthlyRate": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ParkingStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.ParkingType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ParkingStatus": {
            "type": "string",
            "enum": [
                "idle",
                "occupied",
                "faulty"
            ],
            "x-enum-varnames": [
                "Idle",
                "Occupied",
                "Faulty"
            ]
        },
        "models.ParkingType": {
            "type": "string",
            "enum": [
                "permanent",
                "short_term",
                "temporary"
            ],
            "x-enum-varnames": [
                "Permanent",
                "ShortTerm",
                "Temporary"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 JWT 格式的令牌进行授权。格式：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}